% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{togglePlotTransparency}
\alias{togglePlotTransparency}
\title{Toggle plot transparency}
\usage{
togglePlotTransparency()
}
\description{
Toggles on and off plot area transparency on mouse hovering for the Kandle theme.
It is obtained by simply switching between two themes, one in which this is allowed,
one in which this is not.
}
\details{
If Kandle theme is currently not being used, it will automatically switch to it.
}
